Continue if runners fail to remove
The null_resource workaround for unregistering deleted runners from
Github seems to work well, but occasionally can get out of sync with the
VM resources and try to remove runners that no longer exist. My hope is
this is mostly just a result of the churn that happens in a dev
environment, but when it happens it can make it difficult to go ahead
and destroy the module. Adding an on_failure condition to cover this.

Also fmt the module.